revisiting-gnn-curvature-main/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohappyeyeballs=2.4.0=py38haa95532_0
aiohttp=3.10.5=py38h827c3e9_0
aiosignal=1.2.0=pyhd3eb1b0_0
astor=0.8.1=pypi_0
async-timeout=4.0.3=py38haa95532_0
attrdict=2.0.1=pypi_0
attrs=23.1.0=py38haa95532_0
blas=1.0=mkl
blinker=1.8.2=pypi_0
bottleneck=1.3.7=py38h9128911_0
brotli-python=1.0.9=py38hd77b12b_8
ca-certificates=2024.7.2=haa95532_0
certifi=2024.8.30=py38haa95532_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=pypi_0
cloudpickle=3.0.0=pypi_0
colorama=0.4.6=py38haa95532_0
community=1.0.0b1=pypi_0
cuda-cccl=12.6.37=0
cuda-cccl_win-64=12.6.37=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.6.68=0
cuda-opencl-dev=12.6.68=0
cuda-profiler-api=12.6.68=0
cuda-runtime=12.1.0=0
cuda-version=12.6=3
filelock=3.11.0=pypi_0
flask=3.0.3=pypi_0
freetype=2.12.1=ha860e81_0
frozenlist=1.4.0=py38h2bbff1b_0
fsspec=2024.6.1=py38haa95532_0
gmpy2=2.1.2=py38h7f96b67_0
icc_rt=2022.1.0=h6049295_2
idna=3.7=py38haa95532_0
importlib-metadata=7.0.1=py38haa95532_0
importlib_metadata=7.0.1=hd3eb1b0_0
intel-openmp=2023.1.0=h59b6b97_46320
itsdangerous=2.2.0=pypi_0
jinja2=3.1.4=py38haa95532_0
joblib=1.4.2=py38haa95532_0
jpeg=9e=h827c3e9_3
json-tricks=3.17.3=pypi_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.7.68=0
libcurand-dev=10.3.7.68=0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_1
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.1=hd77b12b_0
libuv=1.48.0=h827c3e9_0
libwebp-base=1.3.2=h2bbff1b_0
llvmlite=0.41.0=py38hf2fb9eb_0
lz4-c=1.9.4=h2bbff1b_1
markupsafe=2.1.3=py38h2bbff1b_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py38h2bbff1b_1
mkl_fft=1.3.8=py38h2bbff1b_0
mkl_random=1.2.4=py38h59b6b97_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py38haa95532_0
multidict=6.0.4=py38h2bbff1b_0
networkit=11.0=pypi_0
networkx=3.1=py38haa95532_0
nni=3.0=pypi_0
numba=0.58.1=py38hd77b12b_0
numexpr=2.8.4=py38h7b80656_1
numpy=1.24.3=py38h79a8e48_1
numpy-base=1.24.3=py38h8a87ada_1
nvidia-ml-py=12.560.30=pypi_0
openjpeg=2.5.2=hae555c5_0
openssl=3.0.15=h827c3e9_0
packaging=24.1=py38haa95532_0
pandas=2.0.3=py38h4ed8f06_0
pillow=10.4.0=py38h827c3e9_0
pip=24.2=py38haa95532_0
platformdirs=3.10.0=py38haa95532_0
pooch=1.7.0=py38haa95532_0
pot=0.9.4=pypi_0
prettytable=3.11.0=pypi_0
psutil=5.9.0=py38h2bbff1b_0
pyg=2.4.0=py38_torch_2.1.0_cu121
pyparsing=3.1.2=py38haa95532_0
pysocks=1.7.1=py38haa95532_0
python=3.8.19=h1aa4202_0
python-dateutil=2.9.0post0=py38haa95532_2
python-tzdata=2023.3=pyhd3eb1b0_0
pythonwebhdfs=0.2.3=pypi_0
pytorch=2.1.2=py3.8_cuda12.1_cudnn8_0
pytorch-cuda=12.1=hde6ce7c_5
pytorch-mutex=1.0=cuda
pytz=2024.1=py38haa95532_0
pyyaml=6.0.1=py38h2bbff1b_0
requests=2.32.3=py38haa95532_0
responses=0.25.3=pypi_0
schema=0.7.7=pypi_0
scikit-learn=1.3.0=py38h4ed8f06_1
scipy=1.10.1=py38hdcfc7df_1
setuptools=75.1.0=py38haa95532_0
simplejson=3.19.3=pypi_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h2bbff1b_0
sympy=1.13.2=py38haa95532_0
tbb=2021.8.0=h59b6b97_0
threadpoolctl=3.5.0=py38h9909e9c_0
torchaudio=2.1.2=pypi_0
torchvision=0.15.2=cpu_py38h7187fe4_0
tqdm=4.66.5=py38h9909e9c_0
typeguard=4.1.2=pypi_0
typing_extensions=4.11.0=py38haa95532_0
urllib3=2.2.2=py38haa95532_0
vc=14.40=h2eaa2aa_1
vs2015_runtime=14.40.33807=h98bb1dd_1
wcwidth=0.2.13=pypi_0
websockets=13.1=pypi_0
werkzeug=3.0.4=pypi_0
wheel=0.44.0=py38haa95532_0
win_inet_pton=1.1.0=py38haa95532_0
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
yarl=1.11.0=py38h827c3e9_0
zipp=3.17.0=py38haa95532_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.5=hd43e919_2
